+++ example/lib/inputs/go_password_input.dart
@@ -8,25 +8,49 @@ class GoPasswordInput extends FormFieldModelBase<String>{
 
   @override
   Widget build(BuildContext context, FieldController controller) {
-    return StatefulBuilder(
-      builder: (context,setState) {
-        bool showPassword=false;
-        return RootInput(
-          onChanged: (newValue) => controller.onChange(newValue),
-          initialValue: controller.value,
-          validator: validator,
-          errorText: controller.error,
-          labelText: label,
-          suffixIcon: IconButton(
-            icon: Icon((showPassword)?Icons.visibility:Icons.visibility_off),
-            onPressed: () {
-              showPassword=!showPassword;
-              setState((){});
-            },
-          ),
-          obscureText: showPassword,
-        );
-      }
+    return _PasswordField(
+      controller: controller,
+      label: label,
+      validator: validator,
+    );
+  }
+}
+
+class _PasswordField extends StatefulWidget {
+  final FieldController controller;
+  final String label;
+  final String? Function(String?)? validator;
+
+  const _PasswordField({
+    required this.controller,
+    required this.label,
+    this.validator,
+  });
+
+  @override
+  State<_PasswordField> createState() => _PasswordFieldState();
+}
+
+class _PasswordFieldState extends State<_PasswordField> {
+  bool showPassword = false;
+
+  @override
+  Widget build(BuildContext context) {
+    return RootInput(
+      onChanged: (newValue) => widget.controller.onChange(newValue),
+      initialValue: widget.controller.value,
+      validator: widget.validator,
+      errorText: widget.controller.error,
+      labelText: widget.label,
+      suffixIcon: IconButton(
+        icon: Icon(showPassword ? Icons.visibility : Icons.visibility_off),
+        onPressed: () {
+          setState(() {
+            showPassword = !showPassword;
+          });
+        },
+      ),
+      obscureText: !showPassword,
     );
   }
 }