+++ example/lib/pages/debounce_example_page.dart
@@ -2,6 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:go_form/go_form.dart';
 import 'package:go_form_example/inputs/go_text_input.dart';
 
+import '../inputs/search_input.dart';
+
 /// Example: Debounced Input Field
 ///
 /// This example demonstrates how to use a debounce delay on a text field
@@ -38,21 +40,41 @@ class _DebounceExamplePageState extends State<DebounceExamplePage> {
         title: const Text('Debounce'),
       ),
       body: Padding(
-        padding: const EdgeInsets.symmetric(horizontal: 16,vertical: 8),
+        padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
         child: Column(
           children: [
             DynamicForm(
               fields: [
-                GoTextInput(
+                GoSearchInput(
                   name: 'search',
                   label: 'Search',
                   debounceDuration: const Duration(seconds: 2),
+                  asyncValidator: (v) async {
+                    if (v == null || v.isEmpty) {
+                      return 'Input text';
+                    }
+                    await Future.delayed(const Duration(seconds: 2));
+                    if (v == 'admin') {
+                      return 'Name exist';
+                    }
+                    return null;
+                  },
+                  onDebounceComplete: () async {
+                    print('start onDebounceComplete');
+                    await formController.validateAsync();
+                  },
                 ),
               ],
               controller: formController,
             ),
             const SizedBox(height: 16),
             Text(_output),
+            ElevatedButton(
+              onPressed: () {
+                formController.setError('search', 'error');
+              },
+              child: Text('Add error'),
+            )
           ],
         ),
       ),