+++ lib/src/domain/domain.dart
@@ -1,2 +1,3 @@
 export 'form_field_model_base.dart';
 export 'from_field_data.dart';
+export 'enums/enums.dart';