+++ lib/src/domain/enums/error_reset_mode.dart
@@ -1,0 +1,5 @@
+enum ErrorResetMode {
+  resetOnFocus,
+  resetOnValueChange,
+  noReset,
+}