+++ lib/src/domain/enums/validation_mode.dart
@@ -1,0 +1,5 @@
+enum ValidationMode {
+  valueChange,
+  focusChange,
+  noValidate,
+}